native: avoid import cycle in CryptoEngine
